+++ components/Trade/Advanced/TradingPanel/TradingInput/PlaceOrder/PostTradeDetails.tsx
