+++ components/Trade/Advanced/TradingPanel/TradingInput/index.tsx
@@ -1,2 +1,2 @@
 export { default as PlaceOrder } from './PlaceOrder';
-export { default as AdjustOrder } from './AdjustOrder';
+export { default as ModifyOrder } from './ModifyOrder';
