+++ components/Trade/Advanced/TradingPanel/index.tsx
